added string-typed fields as specified by Nikole and Laura
