+++ scripts/processing/README.md
