+++ scripts/processing/driver.py
@@ -1,5 +1,6 @@
 import map_fields as mapper
 from argparse import ArgumentParser
+import functools
 import json
 import sys
 
@@ -10,29 +11,36 @@ arguments = argparser.parse_args()
 
 try:
     with open(arguments.path) as infile:
-        data = json.load(infile)
+        # Study is an array of json samples with identical fields
+        study = json.load(infile)
 except:
     print('Try better.')
     raise FileExistsError
 
 
-study_sample = data[0]
-ssentries    = study_sample.items()
-processed    = {
-    "airr_compliant": {},
 
-}
+# Reduces known mappings to the "airr_compliant" field for a single sample of a study
+def transform_sample(sample=json):
+    processed    = {
+        "airr_compliant": {},
+    }
+    ssentries    = sample.items()
 
-for entry in ssentries:
-    if entry[0] in mapper.adaptive_to_airr.keys():
-        adaptives_key = entry[0]
-        [processed['airr_compliant'].update(
-            transformed_kvpair) for transformed_kvpair in mapper.adaptive_to_airr[adaptives_key](entry)]
-    else:
-        processed.update([entry])
+    for entry in ssentries:
+        if entry[0] in mapper.adaptive_to_airr.keys():
+            adaptives_key = entry[0]
+            [processed['airr_compliant'].update(
+                transformed_kvpair) for transformed_kvpair in mapper.adaptive_to_airr[adaptives_key](entry)]
+        else:
+            processed.update([entry])
+    return processed
+
+processed_study = [ * map(transform_sample, study) ]
+
+
 
 with open('transformed_first-pass.json', 'w') as outfile:
-    json.dump(processed, outfile)
+    json.dump(processed_study, outfile)
 
 outfile.close()
